requirements.txt
discord.py
cytoolz
requests
